chore: no more LanguagePlugin / internalPlugin / externalPlugin
All we need is a single LanguagePlugin now that there are no internal plugins
