+++ frontend/cli/cmd_new.go
@@ -35,7 +35,7 @@ type newCmd struct {
 // - environment variable overrides
 //
 // Language plugins take time to launch, so we return the one we created so it can be reused in Run().
-func prepareNewCmd(ctx context.Context, k *kong.Kong, args []string) (optionalPlugin optional.Option[languageplugin.LanguagePlugin], err error) {
+func prepareNewCmd(ctx context.Context, k *kong.Kong, args []string) (optionalPlugin optional.Option[*languageplugin.LanguagePlugin], err error) {
 	if len(args) < 2 {
 		return optionalPlugin, nil
 	} else if args[0] != "new" {
